+++ src/features/suggestions/ServingSuggestion.ts
@@ -1,0 +1,8 @@
+import { FoodGroup } from "../../model/Food";
+
+export interface ServingSuggestion {
+  foodName: string;
+  servingSize: string;
+  foodGroup: FoodGroup;
+  bestChoice?: boolean;
+}
