+++ src/features/suggestions/search.ts
@@ -1,6 +1,7 @@
 import Fuse from "fuse.js";
 import _ from "lodash";
 import servings from "./servings";
+import { ServingSuggestion } from "./ServingSuggestion";
 
 const options = {
   // isCaseSensitive: false,
@@ -28,8 +29,8 @@ const searchExpression = (words: string[]) => ({
 
 const match = (words: string[]) => fuse.search(searchExpression(words));
 
-const search = (words: string[], scorePredicate: (res: { score: number; }) => boolean) : any[] => {
-  const res = _.filter(match(words), scorePredicate);
+const search = (words: string[], scorePredicate: (res: { score: number; }) => boolean): { item: ServingSuggestion; score?: number }[] => {
+  const res = _.filter(match(words), scorePredicate) as { item: ServingSuggestion; score?: number }[];
   if (_.size(res) === 0 && _.size(words) > 1) {
     const dropLastWord = () => _.take(words, _.size(words) - 1);
     return search(dropLastWord(), scorePredicate);
